refactor, add logic to harvester
