+++ lib/crawlers/blockHarvester.js
@@ -1,6 +1,6 @@
 // @ts-check
-const {BigNumber} = require('bignumber.js');
-const { shortHash } = require('../utils.js');
+const { BigNumber } = require('bignumber.js');
+const { shortHash, storeExtrinsics } = require('../utils.js');
 
 module.exports = {
   start: async function(api, pool, config) {
@@ -65,21 +65,33 @@ module.exports = {
   harvestBlocks: async function(api, pool, startBlock, endBlock) {
     let addedBlocks = 0;
     while (endBlock >= startBlock) {
+      
       // Start execution
       const startTime = new Date().getTime();
+      
       // Get block hash
       const blockHash = await api.rpc.chain.getBlockHash(endBlock);
+
+      // Get signed block
+      const { block } = await api.rpc.chain.getBlock(blockHash);
+      
       // Get extended block header
       const extendedHeader = await api.derive.chain.getHeader(blockHash);
+      
       // Get block parent hash
       const parentHash = extendedHeader.parentHash;
       
       // Get block extrinsics root
       const extrinsicsRoot = extendedHeader.extrinsicsRoot;
+      
       // Get block state root
       const stateRoot = extendedHeader.stateRoot;
+      
       // Get block events
       const blockEvents = await api.query.system.events.at(blockHash);
+
+      // Store block extrinsics
+      await storeExtrinsics(pool, endBlock, block.extrinsics);
 
       // Get block number finalized
       // TODO: Get finalized from finalitytracker/final_hint extrinsic
