+++ lib/crawlers/blockListener.js
@@ -1,5 +1,5 @@
 // @ts-check
-const { shortHash } = require('../utils.js');
+const { shortHash, storeExtrinsics } = require('../utils.js');
 const {BigNumber} = require('bignumber.js');
 
 module.exports = {
@@ -187,8 +187,8 @@ module.exports = {
           }
         }
 
-        // Store extrinsics
-        await module.exports.storeExtrinsics(pool, blockNumber, block.extrinsics);
+        // Store block extrinsics
+        await storeExtrinsics(pool, blockNumber, block.extrinsics);
 
         // Loop through the Vec<EventRecord>
         await blockEvents.forEach( async (record, index) => {
@@ -231,21 +231,5 @@ module.exports = {
         });
       }
     });
-  },
-  storeExtrinsics: async function (pool, blockNumber, extrinsics) {
-
-    extrinsics.forEach((extrinsic, index) => {
-
-      const signer = extrinsic.signer.toString();
-      const section = extrinsic.toHuman().method.section;
-      const method = extrinsic.toHuman().method.method;
-      const args = JSON.parse(JSON.stringify(extrinsic.args));
-      const hash = extrinsic.hash.toHex();
-      const doc = extrinsic.meta.documentation.toString();
-
-      console.log(`Extrinsic: ${blockNumber}-${index} signer: ${signer} section: ${section} method: ${method} args: ${args} hash: ${hash} doc: ${doc}`)
-
-    });
-    return true
   }
 }
