+++ lib/utils.js
@@ -9,5 +9,54 @@ module.exports = {
     return new Promise(resolve => {
       setTimeout(resolve, ms);
     });
+  },
+  storeExtrinsics: async function (pool, blockNumber, extrinsics) {
+
+    extrinsics.forEach(async (extrinsic, index) => {
+
+      const signer = extrinsic.signer.toString();
+      const section = extrinsic.toHuman().method.section;
+      const method = extrinsic.toHuman().method.method;
+      const args = JSON.parse(JSON.stringify(extrinsic.args));
+      const hash = extrinsic.hash.toHex();
+      const doc = extrinsic.meta.documentation.toString();
+
+      // console.log(`Extrinsic: ${blockNumber}-${index} signer: ${signer} section: ${section} method: ${method} args: ${args} hash: ${hash} doc: ${doc}`);
+      
+      let sql = `SELECT FROM extrinsic WHERE block_number = '${blockNumber}' AND extrinsic_index = '${index}';`;
+      let res = await pool.query(sql);
+
+      if (res.rows.length === 0) {
+      
+        sql = 
+          `INSERT INTO extrinsic (
+            block_number,
+            extrinsic_index,
+            signer,
+            section,
+            method,
+            args,
+            hash,
+            doc
+          ) VALUES (
+            '${blockNumber}',
+            '${index}',
+            '${signer}',
+            '${section}',
+            '${method}',
+            '${args}',
+            '${hash}',
+            '${doc}'
+          );`;
+        try {
+          await pool.query(sql);
+          console.log(`Adding extrinsic ${blockNumber}-${index}`);
+          return true;
+        } catch (error) {
+          console.log(`Error adding extrinsic ${blockNumber}-${index}`);
+          return false;
+        }
+      }
+    });
   }
 };
